% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston_block_groups}
\alias{boston_block_groups}
\title{Boston Census Block Groups}
\format{
An sf object with 646 features and 4 fields:
\describe{
\item{CBG_ID_10}{Census block group ID (12-digit FIPS code, 2010 vintage)}
\item{BG_ID_10}{Block group number within tract}
\item{CT_ID_10}{Parent census tract ID}
\item{NAME}{Block group name}
\item{geometry}{Polygon geometry (EPSG:4326)}
}
}
\source{
U.S. Census Bureau, TIGER/Line Shapefiles 2010, Suffolk County MA.
Downloaded via the tigris R package.
}
\usage{
boston_block_groups
}
\description{
Spatial geometries for Boston census block groups (2010 vintage) for Suffolk
County, Massachusetts. Downloaded from the U.S. Census Bureau TIGER/Line
shapefiles via the tigris package.
}
\examples{
\dontrun{
library(sf)
plot(boston_block_groups["NAME"])
}
}
\keyword{datasets}
