% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boston_tracts}
\alias{boston_tracts}
\title{Boston Census Tracts}
\format{
An sf object with 204 features and 2 fields:
\describe{
\item{CT_ID_10}{Census tract ID (11-digit FIPS code, 2010 vintage)}
\item{NAME}{Census tract name}
\item{geometry}{Polygon geometry (EPSG:4326)}
}
}
\source{
U.S. Census Bureau, TIGER/Line Shapefiles 2010, Suffolk County MA.
Downloaded via the tigris R package.
}
\usage{
boston_tracts
}
\description{
Spatial geometries for Boston census tracts (2010 vintage) for Suffolk
County, Massachusetts. Downloaded from the U.S. Census Bureau TIGER/Line
shapefiles via the tigris package.
}
\examples{
\dontrun{
library(sf)
plot(boston_tracts["NAME"])
}
}
\keyword{datasets}
