% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidybari-package.R
\docType{package}
\name{tidybari-package}
\alias{tidybari}
\alias{tidybari-package}
\title{tidybari: Access Boston Area Research Initiative Data}
\description{
Download and work with data from the Boston Area Research Initiative (BARI) Data Portal, hosted on Harvard Dataverse. Provides tidy access to neighborhood ecometrics, administrative records, and spatial geometries for the City of Boston. Inspired by the tidycensus package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joshrrosen/tidybari}
  \item Report bugs at \url{https://github.com/joshrrosen/tidybari/issues}
}

}
\keyword{internal}
